Code upgrades: * Remove ValidName.java and PersonNameValidator.java * Make ImageEntity.kt field 'checksum' not nullable * Add new fields to ProjectEntity.kt * Change CalendarUnitEntity id type to Int * Improve PersonEntityService.kt, PersonRepository.kt methods * Implement new methods in ProjectEntityService.kt * Upgrade ProjectEntity create script

+++ persistence/src/main/kotlin/io/github/blvckmind/relic/persistence/repository/CalendarUnitEntityRepository.kt
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CalendarUnitEntityRepository : JpaRepository<CalendarUnitEntity, Long> {
+interface CalendarUnitEntityRepository : JpaRepository<CalendarUnitEntity, Int> {
 
     fun getAllBySourceId(sourceId: String, pageable: Pageable): Page<CalendarUnitEntity>
